Simple route mapping
Simple routes (i.e. to HTML files) are defined by a json object. The key
is the route and the value is the path to the file.
A template file must be provided.
The directory in which the HTML files are must be provided.
The routes are built with the buildRoutes function in routes.js.

Added several example routes to index.js.